+++ Customer/src/main/java/com/example/customer/repository/CustomerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CustomerRepository extends JpaRepository<CustomerEntity, Integer> {
+public interface CustomerRepository extends JpaRepository<CustomerEntity, Long> {
     CustomerEntity findByPhone(String phone);
 }
